LARGE refactor for handling all options pt1
Basically make the main option renderer generic and added a bunch of code to configure the individual options and then some more for saving said options.

There's some kind of bug with trying to use the SPW values.

Still needs to update the Dpad options to work with the new setup.