New SX Options Menu System and Options - ENG (#22)
* Add Disable SPW Feature

* Add No SPW option to menu + prep for more options

* Setup - Hardcode initializing look up table in an attempt to save space for code.
* SettingsSet - Swapped order of FNT Messages.
* Render - Replaced JPN ENG branches with reference to lookup table.

Saved space in Render, but that was added to SettingsSet.  Another update could move the lookup table to a FNT message as I would need to anyways for the planned revamped option display.

* Refactor Intro Message Render

Was able to reduce the size of the code by approximately 20% by using more branches for repeated code.

* Use a FNT lookup table for rendering option

Plus a bit more refactoring.

* Option sets to half words + Init Page 2 values

* Update to show the save message on all pages

* Implement Page 2 Options

Probably some refactoring and other things too, but that's all that is obvious with the diff.

* Implement up to 5 options and more pages

* Stack Pointer now being used.

This should free up r14 and r15 for logic and parameters

* LARGE refactor for handling all options pt1

Basically make the main option renderer generic and added a bunch of code to configure the individual options and then some more for saving said options.

There's some kind of bug with trying to use the SPW values.

Still needs to update the Dpad options to work with the new setup.

* Fix SPW Setting Issues - Fixed option changing

Assuming correct values provided, this new version of the options menu should be complete.  Will probably take another pass through for size optimization once changes for SettingSet are finished.

* Added in applying settings on option menu exit Pt1

This section of code will handle the Page 2 options along with the SPW changes sticking correctly.  More work needed, especially for the keys and stage logic.

* Move Default Save Settings check to Render.

This ensures the data is updated after save data is loaded in, automatically updating older save files.

* Fixes for apply Page 2 actions

* Better / correct setting for SPW changes.
* Fixed Last Story and Expert not updating to the correct locations.
* Implemented Unlock Stages and Bosses.
* Implemented Unlock Keys.
* Fixed Remove Keys.

* Update Advertise_EN.fnt

Includes all of the necessary changes for the new menu system.

* Update SX_MCM.txt

Over 77% used space :scream: