FNT file updates - Fix override to Format message  (#21)
* Updated .fnt files

No longer overwrites the Format Memory Card message.

* Set Message Offset to 1 on Credit Shown

This is now required to show the custom message.

* New FNT Redirect code

The code responsible for redirecting the start address for the intro settings has now been updated to also work for the IGT message.  Also, now has a buffer setup for when the window is trying to animate close.

* Add new offsetting based code

Not good enough to prevent the showing the base message.  Will likely pivot to a different type of change.

* Fix for brief incorrect message on offset 0

Found that there was a second memory location that was checked when rendering messages. Setting it along with the main byte seems to help prevent incorrect messages.

* New Comment Symbol for rebase changes

* New Comment Type for FNTOverrideRedirect.asm

* Update SX_MCM.txt

* Update Advertise_EN.fnt